requirements.txt
Pillow>=10.0.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

